Added Working Files and Directories
This is a work in progress and still has a long way to go. 
This is a two tabbed shopping list with the ability to add/remove/delete items and move them from a shopping list to a shopping cart.
There is also a simple sort function within the top menu.
I am currently working on another iteration that implements a considerable amount of functionality, including:
Specialized sorting, multiple threads, multiple lists, a main menu with options and setting preferences, color coded/icon coded grocery types for easy identification and sorting when at the grocery store, syncing with other household members, and more! Stay tuned!